requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Jinja2==3.1.3
Werkzeug==3.0.1 